refactor(ci): share validation helpers
